added trigger param to useclickoutside and unitilized in in the create note component
